+++ print_product_of_two_user_number_inputs.py
@@ -1,8 +1,8 @@
 # B1 Prog04: Create a program that ask user to input 2 numbers. Print the product of the two numbers.
 
 # User Input
-num1 = float(input("First Number: "))
-num2 = float(input("Second Number: "))
+num1 = int(input("First Number: "))
+num2 = int(input("Second Number: "))
 
 # Print product of the two numbers
 prod = num1 * num2
