+++ print_quotient_with_decimal_of_two_user_number_inputs.py
@@ -1,8 +1,8 @@
 # B1 Prog05: Create a program that ask user to input 2 numbers. Print the quotient of the two numbers with the decimal point
 
 # User Input
-num1 = float(input("First Number: "))
-num2 = float(input("Second Number: "))
+num1 = int(input("First Number: "))
+num2 = int(input("Second Number: "))
 
 # Print quotient
 quo = num1 / num2
