+++ print_quotient_with_no_decimal_of_two_user_number_inputs.py
@@ -1,8 +1,8 @@
 # B2 Prog04: Create a program that ask user to input 2 numbers. Print the quotient of the two numbers without the decimal point
 
 # Ask for user input
-num1 = float(input("First Number: "))
-num2 = float(input("Second Number: "))
+num1 = int(input("First Number: "))
+num2 = int(input("Second Number: "))
 
 # Print quotient w/o decimals (whole number only)
 if num2 == 0:
